Remove packages which no longer exist
Removes dent, just and finger-config dependencies.  finger-config was
deprecated long time ago, but the code to install it if finger was not
new enough was still in place.  dent and just were removed because they
were useless.

Do not suggest to install owl.  It's a very bad suggestion.

This also simplifies packaging by using dh7